chore: refactor traces
