--- frontend/src/container/InfraMonitoringK8s/Clusters/ClusterDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import ClusterTraces from './ClusterTraces';
-
-export default ClusterTraces;
