--- frontend/src/container/InfraMonitoringK8s/Deployments/DeploymentDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import DeploymentTraces from './DeploymentTraces';
-
-export default DeploymentTraces;
