--- frontend/src/container/InfraMonitoringK8s/Namespaces/NamespaceDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import NamespaceTraces from './NamespaceTraces';
-
-export default NamespaceTraces;
