--- frontend/src/container/InfraMonitoringK8s/Nodes/NodeDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import NodeTraces from './NodeTraces';
-
-export default NodeTraces;
